TrailsInTheSkyFix: Have code reference TrailsInTheSkyFix NOT TrailsInTheSkyFCFix
Given that this fix works for all three games in
the trilogy it no longer makes sense that the fix
constantly references just the first game. All
nomenclature has been changed to reference the
whole series.

	modified:   CMakeLists.txt
	modified:   README.md
	modified:   install.ps1
	modified:   src/dllmain.cpp

No longer does `install.ps1` have a string for
this file. It is now a standalone file that is
used for copy and pasting.

	new file:   TrailsInTheSkyFix.yml