TrailsInTheSkyFix: Update to v1.1.1
A bug was introduced in v1.1.0 where the camera
zoom in/out feature did not work properly. When
interacting with certain objects and after
cutscenes the camera would glitch out either for
a few seconds or would just break.

This was fixed by simplifying the
tileRenderDistance function so now there are two
hooks in place, 1 where the tile render distance
gets updated to max, or close to max, and another
where the camera zoom in/out feature is
implemented.

Also has extra code for UI centering. This is WIP
and will be deleted in next commit, but kept here
as a placeholder in case the work does get resumed
in the future.

	modified:   src/dllmain.cpp

Updated patternScan to now handle white spaces
as sometimes it would fail to do so and would be
unable to patternScan properly which would lead
to false positives that certain patterns where
not found when they do exist, but the byte array
was computed wrong from the IDA-style string.

	modified:   src/utils.cpp