TrailsInTheSkyFix: Fix `tileRenderFix` to work across all 3 games
Updated `injectHook` function so that mid hooks
are now appended to a static vector declared in
the function. This was needed as in certain
instances of how `injectHook` was called the
midhook would not get injected ie if `injectHook`
was called in a for loop.

	modified:   inc/utils.hpp

Update to `tileRenderFix` signatures. Now this
will work across all 3 games.

	modified:   src/dllmain.cpp

Updated comment above `masterEnable` to also
include features. In the code it effects features
regardless, but the documentation in the yml file
did not indicate this. This is a documentation
enhancement.

	modified:   TrailsInTheSkyFix.yml